FIxing signal bars
Turned out that this is a very common issue in many devices.
The int[] array that is sent from the RADIO to Android is not properly formatted
(meaning that it is not ready for AOSP comsumption).

For XT910 the first two fields of this array contain GSM signal info, and the
rest is garbarge (CDMA and LTE data).

The problem arrises when AOSP, to properly display GSM signal data,
checks if the fields corresponding to LTE data in the array are set to -1,
but Motorola RIL sends them as 0!

Turning on the 'signalstrengh' old ril "fix" this goes way.

More details can be found here:
https://github.com/CyanogenMod/android_frameworks_base/blob/ics/telephony/java/com/android/internal/telephony/RIL.java#L3367
https://github.com/CyanogenMod/android_frameworks_base/blob/ics/telephony/java/android/telephony/SignalStrength.java#L292

Change-Id: Ie2bf5622fab60686c594293edfffba2c0329dc52